requirements.txt
Flask
torch
opencv-python
numpy
Pillow
pytesseract
requests